+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `qobject` module is no longer generated
 - `impl cxx_qt::trait for qobject::T` inside the bridge is now `impl cxx_qt::trait for T`
 - `qobject::T` as the self parameter in the bridge is now `T`
+- `#[cxx_override]`, `#[cxx_final]`, `#[cxx_virtual]` are now independant attributes rather than embedded in `#[qinvokable]`
 
 ### Fixed